+++ samples/ZenUI.Wpf.Gallery/Views/ListBoxView.xaml.cs
@@ -1,0 +1,12 @@
+using System.Windows.Controls;
+
+namespace ZenUI.Wpf.Gallery.Views
+{
+    public partial class ListBoxView : UserControl
+    {
+        public ListBoxView()
+        {
+            InitializeComponent();
+        }
+    }
+}